prepare release and refactoring

+++ gameboy_core/src/mmu/mbc5.rs
@@ -3,9 +3,9 @@ use super::mbc::Mbc;
 
 pub struct Mbc5 {
     cartridge: Cartridge,
-    selected_rom_bank: i32,
-    selected_rom_bank_high: i32,
-    selected_eram_bank: i32,
+    selected_rom_bank: usize,
+    selected_rom_bank_high: usize,
+    selected_eram_bank: usize,
     external_ram_enabled: bool,
     ram_change_callback: Box<dyn FnMut(usize, u8)>,
 }
@@ -19,13 +19,13 @@ impl Mbc for Mbc5 {
             }
             0x4000..=0x7FFF => {
                 let rom = self.cartridge.get_rom();
-                let offset = self.selected_rom_bank as usize * 0x4000;
+                let offset = self.selected_rom_bank * 0x4000;
                 rom[index as usize - 0x4000 + offset]
             }
             0xA000..=0xBFFF => {
                 if self.external_ram_enabled && self.cartridge.get_ram_size() > 0 {
                     let ram = self.cartridge.get_ram();
-                    let offset = self.selected_eram_bank as usize * 0x2000;
+                    let offset = self.selected_eram_bank * 0x2000;
                     ram[index as usize - 0xA000 + offset]
                 } else {
                     0xFF
@@ -39,17 +39,17 @@ impl Mbc for Mbc5 {
         match index {
             0x0000..=0x1FFF => self.external_ram_enabled = (value & 0x0F) == 0x0A,
             0x2000..=0x2FFF => {
-                self.selected_rom_bank = i32::from(value) | (self.selected_rom_bank_high << 8);
+                self.selected_rom_bank = usize::from(value) | (self.selected_rom_bank_high << 8);
                 self.selected_rom_bank &= self.cartridge.get_rom_banks() - 1
             }
             0x3000..=0x3FFF => {
-                self.selected_rom_bank_high = i32::from(value) & 0x01;
+                self.selected_rom_bank_high = usize::from(value) & 0x01;
                 self.selected_rom_bank =
                     (self.selected_rom_bank & 0xFF) | (self.selected_rom_bank_high << 8);
                 self.selected_rom_bank &= self.cartridge.get_rom_banks() - 1
             }
             0x4000..=0x5FFF => {
-                self.selected_eram_bank = i32::from(value) & 0x0F;
+                self.selected_eram_bank = usize::from(value) & 0x0F;
                 self.selected_eram_bank &= self.cartridge.get_ram_banks() - 1;
             }
             0x6000..=0x7FFF => (),
